requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
circuitbreaker==1.4.0
cryptography==41.0.2
discord-py-interactions==5.8.0
discord-typings==0.6.0
discord.py==2.3.1
emoji==2.7.0
exceptiongroup==1.1.2
fastapi==0.100.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
oci==2.108.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
yarl==1.9.2
